# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Katalog Produk Global API",
        "description": "Dokumentasi API untuk Modul 1",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/catalog/categories/{slug}": {
            "get": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Get category detail by slug",
                "description": "Returns the category details",
                "operationId": "getCategoryBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Category Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/v1/catalog/categories/{slug}/breadcrumbs": {
            "get": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Get breadcrumbs for a category",
                "description": "Returns an array of categories representing the path from root to this category",
                "operationId": "getCategoryBreadcrumbs",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Category Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/v1/catalog/products": {
            "get": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Get all active products",
                "description": "Returns paginated active products list",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/catalog/products/{slug}": {
            "get": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Get product detail by slug",
                "description": "Returns full product detail snapshot",
                "operationId": "getProductBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Product Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/v1/catalog/products/id/{id}": {
            "get": {
                "tags": [
                    "Module 1 - Catalog Metadata"
                ],
                "summary": "Get product detail by ID",
                "description": "Returns full product detail snapshot",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
...
}